feat: Implemented infinite scroll for Products page using react-query
This commit refactored the Products page to utilize react-query for
fetching and displaying product data with infinite scroll. The
fetchProductList function from the api.js file was integrated with
the useInfiniteQuery hook. The page now dynamically load and displays
product cards as the user scrolls. Styling and layout adjustments were
made to accommodate the new functionality.

The following changes were make:
- Implemented infinite scroll for Products page
- Refactored Products page to use react-query for data fetching
- Integrated fetchProductList function with usInfiniteQuery hook
- Updated layout and styling to support infinite scroll